Move main method to seperate file
